# openapi.yml
components:
  schemas:
    Graph:
      properties:
        format:
          title: Format
          type: string
        graph:
          title: Graph
          type: string
        id:
          title: Id
          type: string
      required:
      - id
      - graph
      title: Graph
      type: object
    HTTPError:
      properties:
        error:
          title: Error
          type: string
        message:
          title: Message
          type: string
      title: HTTPError
      type: object
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
    ID:
      properties:
        id:
          title: Id
          type: string
      required:
      - id
      title: ID
      type: object
    Message:
      properties:
        message:
          title: Message
          type: string
      title: Message
      type: object
    TemporalID:
      properties:
        id:
          title: Id
          type: string
        timestamp:
          description: Seconds since epoch
          title: Timestamp
          type: integer
      required:
      - id
      title: TemporalID
      type: object
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          title: Location
          type: array
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
      required:
      - loc
      - msg
      - type
      title: ValidationError
      type: object
info:
  description: Historical storage for rdf
  title: rdf-diff-store
  version: 0.1.0
openapi: 3.0.2
paths:
  /api/graphs:
    delete:
      description: Delete graph
      operationId: delete_api_graphs_api_graphs_delete
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ID'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                anyOf:
                - $ref: '#/components/schemas/Message'
                - $ref: '#/components/schemas/HTTPError'
                title: Response Delete Api Graphs Api Graphs Delete
          description: Successful Response
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
          description: Not Found
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
        '500':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
          description: Internal Server Error
      summary: Delete Api Graphs
    get:
      description: Get graph at specific time
      operationId: get_api_graphs_api_graphs_get
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TemporalID'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                anyOf:
                - type: string
                - $ref: '#/components/schemas/Message'
                - $ref: '#/components/schemas/HTTPError'
                title: Response Get Api Graphs Api Graphs Get
          description: Successful Response
        '404':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
          description: Not Found
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
        '500':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
          description: Internal Server Error
      summary: Get Api Graphs
    post:
      description: Store graph
      operationId: post_api_graphs_api_graphs_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Graph'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
        '500':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
          description: Internal Server Error
      summary: Post Api Graphs
  /livez:
    get:
      description: Endpoint for liveness probe.
      operationId: get_livez_livez_get
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
      summary: Get Livez
  /readyz:
    get:
      description: Endpoint for readiness probe.
      operationId: get_readyz_readyz_get
      responses:
        '200':
          content:
            application/json:
              schema: {}
          description: Successful Response
      summary: Get Readyz